Added a .STL file for the heatsink mounting plate.
The M510 requires a mounting plate for the CPU heatsink.  In addition to
helping hold the heatsink in place, the plate also provides mounting
screws for the system's M.2 cards.  This is a quick attempt at modelling
a heat sink plate with places for a few M.2 screws.  This revision
doesn't include holes for the SATA M.2, only the two NVMe M.2s.  I've
added holes for 2260, 2280, and 22110 where possible.

This plate is designed to use heat-set inserts for holding M3 screws
from the heatsink.  I specifically designed for
https://www.mcmaster.com/94180A331/, but any similarly sized insert will
work.  The M.2s should be held in place with M2 screws, probably no
longer than 4mm.